Create search_fields.py from instance.py
